Move dataLookup into src
