+++ src/triggers/TRPR0010.ts
@@ -1,7 +1,7 @@
+import { lookupRemandStatusByCjsCode } from "src/dataLookup"
 import type { Offence } from "src/types/AnnotatedHearingOutcome"
 import { TriggerCode } from "src/types/TriggerCode"
 import type { TriggerGenerator } from "src/types/TriggerGenerator"
-import { lookupRemandStatusByCjsCode } from "src/use-cases/dataLookup"
 
 const triggerCode = TriggerCode.TRPR0010
 const resultCode = 4597
