+++ src/use-cases/enrichHearingOutcome/enrichFunctions/populateBailConditions.ts
@@ -1,5 +1,5 @@
+import { lookupResultQualifierCodeByCjsCode } from "src/dataLookup"
 import type { Result, ResultQualifierVariable } from "src/types/AnnotatedHearingOutcome"
-import { lookupResultQualifierCodeByCjsCode } from "src/use-cases/dataLookup"
 
 const resultQualifierIsBailCondition = (qualifier: ResultQualifierVariable): boolean =>
   qualifier.Code >= "JD" && qualifier.Code <= "JZ"
